Returned specific data types

+++ Code/bmi.h
@@ -2,10 +2,10 @@
 #ifndef BMI_H
 #define BMI_H
 
-double ounces2pounds(double ounces);
-double stones2pounds(double stones);
-double weight2kg(double stones, double pounds, double ounces);
-double height2metres(double feet, double inches);
+double ounces2pounds(int ounces);
+double stones2pounds(int stones);
+double weight2kg(int stones, int pounds, int ounces);
+double height2metres(int feet, int inches);
 char categorise(double kg, double metre);
 
 #endif
